Adds new EpidemicWave model and renames existing one Epidemic. Extracts waves from intervals.
