+++ covid_app/controllers/oc_controller.py
@@ -259,13 +259,14 @@ class OcController(Controller):
 
         analysis = OcWaveAnalysis(test=False)
         print('avg_positive_rates', len(analysis.avg_positive_rates))
-        print('windows', len(analysis.wave.windows))
-        print('phases', len(analysis.wave.phases))
-        print('smooth_phases', len(analysis.wave.smoothed_phases))
+        print('windows', len(analysis.epidemic.windows))
+        print('phases', len(analysis.epidemic.phases))
+        print('smooth_phases', len(analysis.epidemic.smoothed_phases))
 
         #breakpoint()
 
-        pprint(analysis.wave.smoothed_phases)
+        pprint(analysis.epidemic.smoothed_phases)
+        pprint(analysis.epidemic.waves)
         #analysis.windows_to_csv()
         #analysis.export_sample_data_to_csv()
-        print(analysis.export_windows_and_phases_to_csv())
+        #print(analysis.export_windows_and_phases_to_csv())
